==> content/courses/excursions-into-biostatistics/6_Regressions---Correlations-for-the-Advanced.Rmd <==
---
title: Regressions
subtitle: Correlations for the Advanced
date: "2021-02-27"
slug: Regressions - Correlations for the Advanced
categories: [Excursion into Biostatistics]
tags: [R, Statistics]
summary: 'These are exercises and solutions meant as a compendium to my talk on Regression Models.'
authors: [Erik Kusch]
lastmod: '2020-02-27'
featured: no
projects:
output:
  blogdown::html_page:
    keep_md: true
    # toc: true
    # toc_depth: 2
    # number_sections: false
    fig_width: 8
linktitle: Regressions - Correlations for the Advanced
menu:
  Excursions:
    parent: Seminars
    weight: 7
toc: true
type: docs
weight: 7
---

```{r setup, include=FALSE}
rm(list=ls())
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE, tidy='styler', tidy.opts=list(strict=TRUE), fig.width = 15, fig.height = 8)
options(width = 200)
```

## Theory 
These are exercises and solutions meant as a compendium to my talk on Model Selection and Model Building.  

I have prepared some {{< staticref "https://htmlpreview.github.io/?https://github.com/ErikKusch/Homepage/blob/master/static/courses/Excursions-into-Biostatistics/Regressions---Correlations-for-the-Advanced.html" "newtab" >}} Lecture Slides {{< /staticref >}} for this session.

## Our Resarch Project

Today, we are looking at a big (and entirely fictional) data base of the common house sparrow (*Passer domesticus*). In particular, we are interested in the **Evolution of *Passer domesticus* in Response to Climate Change** which was previously explained {{< staticref "courses/excursions-into-biostatistics/research-project/" "newtab" >}} here{{< /staticref >}}.

### The Data

I have created a large data set for this exercise which is available {{< staticref "https://github.com/ErikKusch/Homepage/raw/master/content/courses/excursions-into-biostatistics/Data.rar" "newtab" >}} here{{< /staticref >}} and we previously cleaned up so that is now usable {{< staticref "/courses/excursions-into-biostatistics/data-handling-and-data-assumptions/" "newtab" >}} here{{< /staticref >}}.

### Reading the Data into `R`

Let's start by reading the data into `R` and taking an initial look at it:
```{r}
Sparrows_df <- readRDS(file.path("Data", "SparrowDataClimate.rds"))
head(Sparrows_df)
```

### Hypotheses
Let's remember our hypotheses:  

1. **Sparrow Morphology** is determined by:  
    A. *Climate Conditions* with sparrows in stable, warm environments fairing better than those in colder, less stable ones.  
    B. *Competition* with sparrows in small flocks doing better than those in big flocks.  
    C. *Predation* with sparrows under pressure of predation doing worse than those without.    
2. **Sites**  accurately represent **sparrow morphology**. This may mean:  
    A. *Population status* as inferred through morphology.  
    B. *Site index* as inferred through morphology.  
    C. *Climate* as inferred through morphology.  

Quite obviously, **hypothesis 1** is the only one lending itself well to regression exercises. Since we have three variables that describe sparrow morphology (`Weight`, `Height`, `Wing.Chord`) and multi-response-variable models are definitely above the pay-grade of this material, we need to select one of our morphology variables as our response variable here. Remembering the {{< staticref "courses/excursions-into-biostatistics/classifications-order-from-chaos/" "newtab" >}} Classification exercise{{< /staticref >}}, we recall that `Weight` was the most informative morphological trait so far. Hence, we stick with this one for these exercises.

## `R` Environment

For this exercise, we will need the following packages:

```{r, warning = FALSE, message = FALSE}
install.load.package <- function(x) {
  if (!require(x, character.only = TRUE))
    install.packages(x, repos='http://cran.us.r-project.org')
  require(x, character.only = TRUE)
  }
package_vec <- c(
  "ggplot2", # for visualisation
  "nlme", # for mixed effect models
  "HLMdiag" # for leverage of mixed effect models
)
sapply(package_vec, install.load.package)
```

Using the above function is way more sophisticated than the usual `install.packages()` & `library()` approach since it automatically detects which packages require installing and only install these thus not overwriting already installed packages.

## Linear Regression

Remember the **Assumptions of Linear Regression**:  

1. Variable values follow homoscedasticity (equal variance across entire data range)    
2. Residuals follow normal distribution (normality)    
3. Absence of influential outliers    
4. Response and Predictor are related in a linear fashion    

### Climate Conditions

#### `Weight` as a result of average temperature (`TAvg`)  
Before we begin, let's plot the data we want to model:
```{r}
ggplot(data = Sparrows_df, aes(y = Weight, x = TAvg)) +
  stat_smooth(method = "lm") + geom_point() + 
  theme_bw()
```
I have an inkling that we might run into some issues here, but let's continue for now:

Let's build the actual model:
```{r}
H1_ClimateTavg <- lm(Weight ~ TAvg, data = Sparrows_df)
par(mfrow=c(2,2))
plot(H1_ClimateTavg)
```
While the meeting of most of the assumptions here might be debatable, we certainly cannot accept a linear model with residuals this non-normal distributed. Sow hat do we? We try to remove as many confounding effects as possible!

Remember the plot-locations, climates, and population statues in the data set (go back {{< staticref "courses/excursions-into-biostatistics/research-project/" "newtab" >}} here{{< /staticref >}} if necessary). How about we look exclusively at stations in the Americas which are all housing introduced sparrow populations and almost exclusively lie in coastal habitats? I'll remove all non-coastal climate sites and only consider the central and North America here. Let's do that:

```{r}
# everything west of -7° is on the Americas
# every that's coastal climate type is retained
# everything north of 11° is central and north America
CentralNorthAm_df <- Sparrows_df[Sparrows_df$Longitude < -7 & Sparrows_df$Climate == "Coastal" & Sparrows_df$Latitude > 11 , ]
ggplot(data = CentralNorthAm_df, aes(y = Weight, x = TAvg)) +
  stat_smooth(method = "lm") + geom_point() + 
  theme_bw()
H1_ClimateTavg <- lm(Weight ~ TAvg, data = CentralNorthAm_df)
par(mfrow=c(2,2))
plot(H1_ClimateTavg)
```
This looks sensible to me! The scatterplot shows that sparrows are lighter in warmer areas which makes sense to me.

```{r}
summary(H1_ClimateTavg)
```
Our model estimates show the same pattern.

#### `Weight` as a result of temperature variability (`TSD`)
I'll continue with my North American, coastal subset here:
```{r}
# everything west of -7° is on the Americas
# every that's coastal climate type is retained
# everything north of 11° is central and north America
CentralNorthAm_df <- Sparrows_df[Sparrows_df$Longitude < -7 & Sparrows_df$Climate == "Coastal" & Sparrows_df$Latitude > 11 , ]
ggplot(data = CentralNorthAm_df, aes(y = Weight, x = TSD)) +
  stat_smooth(method = "lm") + geom_point() + 
  theme_bw()
H1_ClimateTSD <- lm(Weight ~ TSD, data = CentralNorthAm_df)
par(mfrow=c(2,2))
plot(H1_ClimateTSD)
summary(H1_ClimateTSD)
```
Again all assumptions are met and the model itself is very intuitive: The more variable the climate, the heavier the sparrows.

#### `Weight` as a result of both temperature mean (`TAvg`) and temperature variability (`TSD`)  

Naturally, we continue with the same subset of the data as before:
```{r}
# everything west of -7° is on the Americas
# every that's coastal climate type is retained
# everything north of 11° is central and north America
CentralNorthAm_df <- Sparrows_df[Sparrows_df$Longitude < -7 & Sparrows_df$Climate == "Coastal" & Sparrows_df$Latitude > 11 , ]
H1_ClimateCont <- lm(Weight ~ TAvg + TSD, data = CentralNorthAm_df)
par(mfrow=c(2,2))
plot(H1_ClimateCont)
summary(H1_ClimateCont)
```

Interestingly, average temperature has a different (and weaker) effect now than when investigated in isolation. Variability of temperature has become even more important (i.e. stronger effect). So which model should we use? That's exactly what we'll investigate in our session on {{< staticref "courses/excursions-into-biostatistics/excursion-into-biostatistics/" "newtab" >}} Model Selection and Validation{{< /staticref >}}!

### Competition
Next, we build models which aim to explain sparrow `Weight` through variables pertaining to competition. To do so, I first calculate the size for each flock at each site and append these to each bird:
```{r}
FlockSizes <- with(Sparrows_df, table(Flock, Index))
Sparrows_df$Flock.Size <- NA
for(Flock_Iter in rownames(FlockSizes)){ # loop over flocks
  for(Site_Iter in colnames(FlockSizes)){ # loop over sites
    Positions <- Sparrows_df$Index == Site_Iter & Sparrows_df$Flock == Flock_Iter
    Sparrows_df$Flock.Size[Positions] <- FlockSizes[Flock_Iter,Site_Iter]
  }
}
```
With that done, we can now build our models up again like we did with the climate data before.

#### `Weight` as a result of Home Range size (`Home.Range`)  
```{r}
ggplot(data = Sparrows_df, aes(x = Home.Range, y = Weight)) +
  geom_boxplot() + theme_bw()
H1_CompetitionHR <- lm(Weight ~ Home.Range, data = Sparrows_df)
par(mfrow=c(2,2))
plot(H1_CompetitionHR)
```
Nope. Those residuals have me nope out. We won't use this model.

#### `Weight` as a result of Flock Size size (`Flock.Size`)  
```{r}
ggplot(data = Sparrows_df, aes(x = Flock.Size, y = Weight)) +
  stat_smooth(method = "lm") + geom_point() + theme_bw()
H1_CompetitionFS <- lm(Weight ~ Flock.Size, data = Sparrows_df)
par(mfrow=c(2,2))
plot(H1_CompetitionFS)
summary(H1_CompetitionFS)
```
Now that's a neat model! Not only do the diagnostics plot look spot-on (not showing these here), the relationship is strong and clear to see - the bigger the flock, the lighter the sparrow.

#### `Weight` as a result of Home Range size (`Home.Range`) and Flock Size size (`Flock.Size`)  
```{r}
ggplot(data = Sparrows_df, aes(x = Flock.Size, y = Weight, col = Home.Range)) +
  geom_point() + stat_smooth(method = "lm") + theme_bw()
H1_CompetitionFULL <- lm(Weight ~ Home.Range * Flock.Size, data = Sparrows_df)
par(mfrow=c(2,2))
plot(H1_CompetitionFULL)
summary(H1_CompetitionFULL)
```

We see that our model is unsure of what to do with the home ranges when it comes to medium-sized ranges. The scatterplot shows that is is probably due to us not having a lot of samples for medium-sized home-ranges. Aside from that, our model meets all assumptions and produces quite intuitive parameter estimates.

### Predation
Next, we look at sparrow `Weight` through the lens of predation. To do so, we need to recode all `NA`s in the `Predator.Type` variable into something else for our models ro tun properly. I chose `"None"` here to indicate that there is no predation-pressure at these sites:
```{r}
levels(Sparrows_df$Predator.Type) <- c(levels(Sparrows_df$Predator.Type), "None")
Sparrows_df$Predator.Type[is.na(Sparrows_df$Predator.Type)] <- "None"
```
With that taken care of, we again build our models one-by-one.

Again, because of issues with normality of residuals, we default to our three sites across Central and North America, which are of coastal climate:
```{r}
# everything west of -7° is on the Americas
# every that's coastal climate type is retained
# everything north of 11° is central and north America
CentralNorthAm_df <- Sparrows_df[Sparrows_df$Longitude < -7 & Sparrows_df$Climate == "Coastal" & Sparrows_df$Latitude > 11 , ]
```

#### `Weight` as a result of `Predator.Presence`  
```{r}
ggplot(data = CentralNorthAm_df, aes(x = Predator.Presence, y = Weight)) +
  geom_boxplot() + theme_bw()
H1_PredationPresence <- lm(Weight ~ Predator.Presence, data = CentralNorthAm_df)
par(mfrow=c(2,2))
plot(H1_PredationPresence)
summary(H1_PredationPresence)
```

According to our model, sparrows under pressure of predation are lighter than those which aren't. Does that make sense? Intuitively, yes, but I would argue that there are too many confounding variable here to be sure.

#### `Weight` as a result of `Predator.Type`  
```{r}
ggplot(data = CentralNorthAm_df, aes(x = Predator.Type, y = Weight)) +
  geom_boxplot() + theme_bw()
H1_PredationType <- lm(Weight ~ Predator.Type, data = CentralNorthAm_df)
par(mfrow = c(2,2))
plot(H1_PredationType)
summary(H1_PredationType)
```
OK. This clearly shows that there are either some confounds present or that our data shows something very counter-intuitive. Sparrows under nor predation are lighter than sparrows under non-avian predation? That makes no sense to me.

#### `Weight` as a result of `Predator.Presence` and `Predator.Type`  
```{r}
ggplot(data = CentralNorthAm_df, aes(x = Predator.Presence, y = Weight, fill = Predator.Type)) +
  geom_boxplot() + theme_bw()
H1_PredationFULL <- lm(Weight ~ Predator.Presence + Predator.Type, data = Sparrows_df)
par(mfrow = c(2,2))
plot(H1_PredationFULL)
summary(H1_PredationFULL)
```
Those residuals don't look good, but that's not what I am after with this model. Immediately, we should notice that our model returns `NA` for the parameter estimate of `Predator.TypeNone`. Why does that happen? Because `Predator.TypeNone` coincides with `Predator.PresenceNo` (the `Intercept`) of this model and so does not provide any additional information. In fact, `Predator.Type` offers all the information of `Predator.Presence` and then some! Consequently, including both in a model does not make sense and we can immediately disqualify this model.

### Null & Full Model
For some comparison further down the line, we need a null model:
```{r}
H1_Null_Sparrows <- lm(Weight ~ 1, data = Sparrows_df)
H1_Null_CNA <- lm(Weight ~ 1, data = CentralNorthAm_df)
```

Our full model contains all of our aforementioned variables/parameters to the best of our knowledge/intuition at this point. We will get to making this model better later. Don't worry: 
```{r}
H1_FULL_Sparrows <- lm(Weight ~ Climate + TAvg + TSD + Home.Range * Flock.Size + Predator.Type, data = Sparrows_df)
par(mfrow = c(2,2))
plot(H1_FULL_Sparrows)
summary(H1_FULL_Sparrows)
```
Already at this point, it is interesting to point out how some previously non-significant effects drop out while other become significant due to the inclusion of all parameters in one model.

Now we also need a full model for our Central-/North-America data:
```{r}
H1_FULL_CNA <- lm(Weight ~ TAvg + TSD + Home.Range * Flock.Size + Predator.Type, data = CentralNorthAm_df)
par(mfrow = c(2,2))
plot(H1_FULL_CNA)
summary(H1_FULL_CNA)
```

### Summary of Linear Regression
To streamline the next few steps of our exercise on {{< staticref "courses/excursions-into-biostatistics/excursion-into-biostatistics/" "newtab" >}} Model Selection and Validation{{< /staticref >}}, we combine all of our models into a `list` object for now:
```{r}
H1_ModelSparrows_ls <- list(H1_Null_Sparrows,
                            H1_CompetitionFS,
                            H1_CompetitionFULL,
                            H1_FULL_Sparrows)
names(H1_ModelSparrows_ls) <- c("Null", 
                                "Comp_Flock.Size", "Comp_Full",
                                "Full")
H1_ModelCNA_ls <- list(H1_Null_CNA,
                       H1_ClimateTavg,
                       H1_ClimateTSD,
                       H1_ClimateCont, 
                       H1_PredationPresence,
                       H1_PredationType,
                       H1_FULL_CNA)
names(H1_ModelCNA_ls) <- c("Null", 
                           "Clim_TAvg", "Clim_TSD", "Clim_Full",
                           "Pred_Pres", "Pred_Type",
                           "Full")
```

## Mixed Effect Models

Remember the **Assumptions of Mixed Effect Models**:  

1. Variable values follow homoscedasticity (equal variance across entire data range)    
2. Residuals follow normal distribution (normality)    
3. Absence of influential outliers    
4. Response and Predictor are related in a linear fashion    

### Climate Conditions
Here, I create a mixed effect model that is accounting predicting `Weight` with a combination `TAvg` and `TSD` while accounting for random intercepts of `Index` and `Population.Status`. As we learned in our exercise on {{< staticref "courses/excursions-into-biostatistics/classifications-order-from-chaos/" "newtab" >}} Classifications{{< /staticref >}}, these two are probably important in controlling for some morphology effects in `Weight`.


I create such a model for:

1. The **entire sparrow data data set**  
2. Only the **Central-/North-America sparrows**  

to make models comparable to our previous basic, linear models.

#### Global Model
```{r}
H1_Climate_ME_Sparrows <- lme(Weight ~ TAvg + TSD,
                     random = list(Population.Status = ~1, 
                                   Index = ~1), 
                     data = Sparrows_df)
par(mfrow=c(2,2))
plot(fitted(H1_Climate_ME_Sparrows), resid(H1_Climate_ME_Sparrows)) #Homogeneity of Variances, values around 0, no pattern --> good!
plot(fitted(H1_Climate_ME_Sparrows), Sparrows_df$Weight) # Linearity, there is a clear pattern here --> bad!
qqnorm(resid(H1_Climate_ME_Sparrows)) # Normality, residuals are normal distributed -> good!
hist(leverage(H1_Climate_ME_Sparrows)[,1]) # Leverage, not really any outliers --> good!
```
Quite evidently, the assumption of linearity is violated. Notice that we would not like to use this model for predictions, but will carry it forward for the purpose of subsequent {{< staticref "courses/excursions-into-biostatistics/excursion-into-biostatistics/" "newtab" >}} Model Selection and Validation{{< /staticref >}}.

```{r}
summary(H1_Climate_ME_Sparrows)
```


#### Central-/North-American Model
Accounting for population status in among the Central-/North-American sites makes no sense as all of them are introduced populations.
```{r}
H1_Climate_ME_CNA <- lme(Weight ~ TAvg + TSD,
                     random = list(Index = ~1), 
                     data = CentralNorthAm_df)
par(mfrow=c(2,2))
plot(fitted(H1_Climate_ME_CNA), resid(H1_Climate_ME_CNA)) #Homogeneity of Variances, values around 0, no pattern --> good!
plot(fitted(H1_Climate_ME_CNA), CentralNorthAm_df$Weight) # Linearity, there is no clear pattern here --> good!
qqnorm(resid(H1_Climate_ME_CNA)) # Normality, residuals are normal distributed -> good!
hist(leverage(H1_Climate_ME_CNA)[,1]) # Leverage, not really any outliers --> good!
```

With this model, all of our assumption are met nicely!

```{r}
summary(H1_Climate_ME_CNA)
```

### Competition

With competition effects as my goal, I can now include climate classification as one of the random effects! Of course, this is with the exception of the Central-/North-American data as all of these sites are of the type "Coastal".

Again, we do so for a global and a local model:

#### Global Model
```{r}
H1_Comp_ME_Sparrows <- lme(Weight ~ Home.Range * Flock.Size,
                     random = list(Population.Status = ~1, 
                                   Climate = ~1), 
                     data = Sparrows_df)
par(mfrow=c(2,2))
plot(fitted(H1_Comp_ME_Sparrows), resid(H1_Comp_ME_Sparrows)) #Homogeneity of Variances, values around 0, no pattern --> good!
plot(fitted(H1_Comp_ME_Sparrows), Sparrows_df$Weight) # Linearity, there is a clear pattern here --> bad!
qqnorm(resid(H1_Comp_ME_Sparrows)) # Normality, residuals are normal distributed -> good!
hist(leverage(H1_Comp_ME_Sparrows)[,1]) # Leverage, not really any outliers --> good!
```

#### Central-/North-American Model
```{r}
H1_Comp_ME_CNA <- lme(Weight ~ Home.Range * Flock.Size,
                     random = list(Climate = ~1), 
                     data = CentralNorthAm_df)
par(mfrow=c(2,2))
plot(fitted(H1_Comp_ME_CNA), resid(H1_Comp_ME_CNA)) #Homogeneity of Variances, values around 0, no pattern --> good!
plot(fitted(H1_Comp_ME_CNA), CentralNorthAm_df$Weight) # Linearity, there is no clear pattern here --> good!
qqnorm(resid(H1_Comp_ME_CNA)) # Normality, residuals are normal distributed -> good!
hist(leverage(H1_Comp_ME_CNA)[,1]) # Leverage, not really any outliers --> good!
```

### Predation

With predation effects, I do the same as with competition effects and set a random intercept for climate classification at each site for a global and a local model: 

#### Global Model

```{r}
H1_Pred_ME_Sparrows <- lme(Weight ~ Predator.Type,
                     random = list(Population.Status = ~1, 
                                   Climate = ~1), 
                     data = Sparrows_df)
par(mfrow=c(2,2))
plot(fitted(H1_Pred_ME_Sparrows), resid(H1_Pred_ME_Sparrows)) #Homogeneity of Variances, values around 0, no pattern --> good!
plot(fitted(H1_Pred_ME_Sparrows), Sparrows_df$Weight) # Linearity, there is a clear pattern here --> bad!
qqnorm(resid(H1_Pred_ME_Sparrows)) # Normality, residuals are normal distributed -> good!
hist(leverage(H1_Pred_ME_Sparrows)[,1]) # Leverage, not really any outliers --> good!
```

#### Central-/North-American Model
```{r}
H1_Pred_ME_CNA <- lme(Weight ~ Predator.Type,
                     random = list(Index = ~1), 
                     data = CentralNorthAm_df)
par(mfrow=c(2,2))
plot(fitted(H1_Pred_ME_CNA), resid(H1_Pred_ME_CNA)) #Homogeneity of Variances, values around 0, no pattern --> good!
plot(fitted(H1_Pred_ME_CNA), CentralNorthAm_df$Weight) # Linearity, there is no clear pattern here --> good!
qqnorm(resid(H1_Pred_ME_CNA)) # Normality, residuals are normal distributed -> good!
hist(leverage(H1_Pred_ME_CNA)[,1]) # Leverage, not really any outliers --> good!
```

### Full Model
#### Global Model
```{r}
H1_Full_ME_Sparrows <- lme(Weight ~ Predator.Type + Flock.Size*Home.Range + TAvg + TSD,
                     random = list(Population.Status = ~1), 
                     data = Sparrows_df)
par(mfrow=c(2,2))
plot(fitted(H1_Full_ME_Sparrows), resid(H1_Full_ME_Sparrows)) #Homogeneity of Variances, values around 0, no pattern --> good!
plot(fitted(H1_Full_ME_Sparrows), Sparrows_df$Weight) # Linearity, there is a clear pattern here --> bad!
qqnorm(resid(H1_Full_ME_Sparrows)) # Normality, residuals are normal distributed -> good!
hist(leverage(H1_Full_ME_Sparrows)[,1]) # Leverage, not really any outliers --> good!
```

#### Central-/North-American Model
```{r}
H1_Full_ME_CNA <- lme(Weight ~ Flock.Size*Home.Range + TAvg + TSD,
                     random = list(Index = ~1), 
                     data = CentralNorthAm_df)
par(mfrow=c(2,2))
plot(fitted(H1_Full_ME_CNA), resid(H1_Full_ME_CNA)) #Homogeneity of Variances, values around 0, no pattern --> good!
plot(fitted(H1_Full_ME_CNA), CentralNorthAm_df$Weight) # Linearity, there is no clear pattern here --> good!
qqnorm(resid(H1_Full_ME_CNA)) # Normality, residuals are normal distributed -> good!
hist(leverage(H1_Full_ME_CNA)[,1]) # Leverage, not really any outliers --> good!
```

### Summary of Mixed Effect Models
Mixed effect models are hard and I have certainly not given them full credit for what they are worth here. I highly suggest giving [this](https://ourcodingclub.github.io/tutorials/mixed-models/#three) a read if you see yourself using mixed effect models.

For now, I want to carry along only the full models for our session on {{< staticref "courses/excursions-into-biostatistics/excursion-into-biostatistics/" "newtab" >}} Model Selection and Validation{{< /staticref >}}:

```{r}
H1_ModelSparrows_ls$Mixed_Full <- H1_Full_ME_Sparrows
H1_ModelCNA_ls$Mixed_Full <- H1_Full_ME_CNA
```

## Generalised Linear Models

For a generalised linear model, we may want to run a logistic regression, which we {{< staticref "courses/excursions-into-biostatistics/classifications-order-from-chaos/#binary-logistic-regression-1" "newtab" >}} already did{{< /staticref >}}. 

For a different example, we now turn to *poisson models* by trying to understand how `Flock.Size` comes about.

Firstly, we limit our data set to necessary variables and then cut all duplicate rows:
```{r}
Flock_df <- Sparrows_df[,c("Flock.Size", "TAvg", "TSD", "Index", "Climate", "Predator.Type")]
Flock_df <- unique(Flock_df)
```

Now we are ready to build a basic linear model:
```{r}
Flock_lm <- lm(Flock.Size ~ TAvg * TSD * Predator.Type, data = Flock_df)
par(mfrow = c(2,2))
plot(Flock_lm)
```

Everything looks line but that Scale-Location plot. That one seems to indicate that variance in our `Flock.Size` increases as the `Flock.Size` itself increases - a very typical example of a poisson distributed variable.

GLMs to the rescue with the poisson GLM:

```{r}
poisson.model <- glm(Flock.Size ~ TAvg * TSD * Predator.Type, data = Flock_df, family = poisson(link = "log"))
par(mfrow=c(2,2))
plot(poisson.model)
```
We fixed it! Now just for the parameter estimates:
```{r}
summary(poisson.model)
```

## Final Models
In our upcoming {{< staticref "courses/excursions-into-biostatistics/excursion-into-biostatistics/" "newtab" >}} Model Selection and Validation{{< /staticref >}} Session, we will look into how to compare and validate models. We now need to select some models we have created here today and want to carry forward to said session.

We have already created `list` objects for this purpose. Let's save them alongside the data that was used to create them (in the case of the localised models, at least). Let's save these as a separate object ready to be loaded into our `R` environment in the coming session: 

```{r}
save(H1_ModelSparrows_ls, H1_ModelCNA_ls, CentralNorthAm_df, Sparrows_df, file = file.path("Data", "H1_Models.RData"))
```

## SessionInfo
```{r}
sessionInfo()
```
